+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2020-09-12
+
+## Added
+
+- Added Configuration properties:  `markyMarkdown.edit.emphasisCharacter` and `markyMarkdown.edit.unorderedListCharacter` to specify preferences for markdown characters used.
+- Added a check when removing an unordered list to see if it is a thematic break (horizontal rule).
+
+### Changed
+
+- Changed the title of command `marky-edit.toggleHorizontalRule` to include the text "thematic break".
+
+### Fixed
+
+- Fixed command title for unordered and ordered lists.
+- Tests needed to be updated with correct extension name for commands.
+- For ordered and unordered lists, when there is no selection, the list inserted would create a list item that could not be edited without deleting the accompanying markdown!
+
 ## [0.1.0] - 2020-09-09
 
 ### Added